Mock login and sessions created
You are now capable of logging in if you are a user in the database.
DBRow and expressServer changes to reflect this. New util objects
Compare and CookieValidator are both helpers to validate if sessions
are active or not. Login page created and is automatically routed to.
Cookies implemented to preserve sessions in the browser. Tables changed
to reflect session object properly. Client js written to communicate
with the server.