Advanced and Settings Pages
Pushing in changes so that others can begin to look and give feedback.
The Settings page allows users to interact with their user settings.
The advanced page allows for advanced search queries which will
(hopefully) tie in with new search features I implement for DBRow in
the next couple weeks.